[Editor 2D] Resolve bug about deleting item/connector/connection system

+++ src/javascript/DeleteConnectionGUI.js
@@ -17,30 +17,32 @@ function deleteConnection(connection, canvas, connections, draw2DConnections, de
 	}
 	var documentsToDelete = documentsToDelete.filter(function(value, index, self) { return self.indexOf(value) === index; }); // permet les id "doublons" dans un tableau
 	var connectorsToDelete = connectorsToDelete.filter(function(value, index, self) { return self.indexOf(value) === index; }); // permet les id "doublons" dans un tableau
-	for (var i = 0; i < documentsToDelete.length; i++) {
-		deleteDocument(documentsToDelete[i]);
-	}
-	// suppression des "other" connectors [otherConnectorsToDelete]
-	console.log(connectorsToDelete.length);
-	for (var i = 0; i < connectorsToDelete.length; i++) {
-		var params = connectorsToDelete[i].split(";"); // params[0] => connector id, params[1] => connector type
-		var connectorId = params[0];
-		var connectorType = params[1];
-		console.log(connectorId);
-		console.log(connectorType);
-		switch (connectorType) {
-			case "output":
-				var allConnections = JSON.parse($.get('https://diagroo.couchappy.com/diagroo/_design/connection/_view/getConnectionByOutputConnector', {'key': '"' + connectorId + '"'}).responseText).rows;
-				if (allConnections.length == 0) {
-					deleteDocument(connectorId);
-				}
-				break;
-			case "input":
-				var allConnections = JSON.parse($.get('https://diagroo.couchappy.com/diagroo/_design/connection/_view/getConnectionByInputConnector', {'key': '"' + connectorId + '"'}).responseText).rows;
-				if (allConnections.length == 0) {
-					deleteDocument(connectorId);
-				}
-				break;
+	if (deleteOnDatabase) {
+		for (var i = 0; i < documentsToDelete.length; i++) {
+			deleteDocument(documentsToDelete[i]);
+		}
+		// suppression des "other" connectors [otherConnectorsToDelete]
+		console.log(connectorsToDelete.length);
+		for (var i = 0; i < connectorsToDelete.length; i++) {
+			var params = connectorsToDelete[i].split(";"); // params[0] => connector id, params[1] => connector type
+			var connectorId = params[0];
+			var connectorType = params[1];
+			console.log(connectorId);
+			console.log(connectorType);
+			switch (connectorType) {
+				case "output":
+					var allConnections = JSON.parse($.get('https://diagroo.couchappy.com/diagroo/_design/connection/_view/getConnectionByOutputConnector', {'key': '"' + connectorId + '"'}).responseText).rows;
+					if (allConnections.length == 0) {
+						deleteDocument(connectorId);
+					}
+					break;
+				case "input":
+					var allConnections = JSON.parse($.get('https://diagroo.couchappy.com/diagroo/_design/connection/_view/getConnectionByInputConnector', {'key': '"' + connectorId + '"'}).responseText).rows;
+					if (allConnections.length == 0) {
+						deleteDocument(connectorId);
+					}
+					break;
+			}
 		}
 	}
 }
